logging: add ExecutionError type and unify execution error handling
Add internal/logging/execution_error.go implementing ExecutionError with Error and Unwrap methods.

Update cmd/runner/main.go to use logging.ExecutionError and pass the structured execution error to the handler.

Refactor internal/logging/pre_execution_error.go: extract handleErrorCommon to consolidate shared logging/output behavior and update handlers to call it.

Adjust tests in internal/logging to construct ExecutionError and call the new HandleExecutionError signature.
